requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
cryptography==42.0.7
Django==4.2.11
django-crispy-forms==2.1
django-extensions==3.2.3
django-sslserver==0.22
djangorestframework==3.15.1
idna==3.7
MarkupSafe==2.1.5
pycparser==2.22
pyOpenSSL==24.1.0
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
